Add support for AmRectangularChart for API, JSO, JSNI and JSON Refactor design for event processing across dom and non-dom AmChartEvent and AmChartDomEvent. (Pending)

+++ src/com/amcharts/api/IsAmChart.java
@@ -73,9 +73,9 @@ public interface IsAmChart
 
 	void setHandDrawThickness( double handDrawThickness );
 
-	Integer getHideBalloonTime();
+	double getHideBalloonTime();
 
-	void setHideBalloonTime( Integer hideBalloonTime );
+	void setHideBalloonTime( double hideBalloonTime );
 
 	String getLanguage();
